apriori-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
mysql-connector-python==8.3.0
pandas==2.2.0
mlxtend==0.23.1
python-dotenv==1.0.0